requirements.txt
flask
flask-cors
pandas
numpy
gunicorn
gtts
markdown
beautifulsoup4
faiss-cpu
transformers
sentence-transformers
python-pptx
concurrent-log-handler
Pillow
nltk
torch
psycopg2-binary
python-dotenv
Flask-Login
Werkzeug
duckdb
google-generativeai
openpyxl
tenacity
scipy
